renamed parser binary to caraparse
