--- carapace/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/rsteube/carapace/carapace/cmd"
-
-func main() {
-	cmd.Execute()
-}
