+++ caraparse/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/rsteube/carapace/caraparse/cmd"
+
+func main() {
+	cmd.Execute()
+}
